feat(simulation): Added NFC block demo code for Trevor
Added NFC block demo code for Trevor. This allows a block to send its id and location to the simulator, which will display the blocks location. Blocks can be added and edited (moved to a new location). Remember to check that the IP address is the IP of the simulator.

Ref: