feat(all): Added blueprint factory
Added factory to generate blueprints (to remove creating blueprints directly within each module). Several blueprints are predefined, such as the playground, a pyramid shape, cube shape, random shape (still needs some work as some blocks will be floating), planar shape, and a "BlockWorld" which is meant for the robot to test climbing. This code should be placed in the shared library. The code within all of the components should also be refactored. These changes have not yet been tested with the structure module.

Ref: