fix(robot): Robot no longer stepping on itself
BREAKING CHANGE: Robot is no longer stepping on itself. The base is no longer correct however. The robot base utility has been completely removed, and so now the only bases are the a and d link positions. Also created an exception to check if the end effectors are at the same location.

Ref: